requirements.txt
altgraph==0.17.4
click==8.1.7
DateTime==5.4
packaging==23.2
pyinstaller==6.4.0
pyinstaller-hooks-contrib==2024.1
PyQt6==6.4.2
pyqt6-plugins==6.4.2.2.3
PyQt6-Qt6==6.4.3
PyQt6-sip==13.6.0
pyqt6-tools==6.4.2.3.3
pyserial==3.5
python-dotenv==1.0.1
pytz==2024.1
qt6-applications==6.4.3.2.3
qt6-tools==6.4.3.1.3
zope.interface==6.2
